- Solved Transition Issues (#4654)
